server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
influxdb-client
python-logstash-async 